+++ com.unity.shadergraph/Editor/Generation/ShaderGraphVfxAsset.cs
@@ -35,6 +35,9 @@ namespace UnityEditor.ShaderGraph.Internal
         public bool lit;
 
         [SerializeField]
+        public bool alphaClipping;
+
+        [SerializeField]
         internal GraphCompilationResult compilationResult;
 
         [SerializeField]